# services/sql-assessment-service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SQL Assessment Service API",
    "version": "1.0.0",
    "description": "REST API for SQL exercise generation, student query grading, and natural-language description of SQL queries in an educational context.",
    "contact": {
      "name": "ALADIN Project"
    }
  },
...
  "paths": {
    "/api/database/analyze-database": {
      "post": {
        "summary": "Analyze and register a PostgreSQL database",
        "description": "Connects to a PostgreSQL database, extracts its full schema metadata (tables, columns, FK relationships, cardinalities), and stores it in the in-memory registry under a generated database key derived from host, port, and schema. The key is required by all subsequent endpoints.\n",
        "tags": [
          "Database"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeDatabaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Database successfully analyzed and registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid connection info or connection failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Schema extraction failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error description (localised).",
            "example": "Unable to connect to the database."
          },
          "code": {
            "type": "string",
            "description": "Optional machine-readable error code.",
            "example": "NON_SELECT"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Database successfully analyzed."
          }
        }
      },
      "PostgresConnectionInfo": {
        "type": "object",
        "required": [
          "type",
          "host",
          "port",
          "username",
          "password",
          "database",
          "schema"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "postgres"
            ],
            "example": "postgres"
          },
          "host": {
            "type": "string",
            "example": "localhost"
          },
          "port": {
            "type": "number",
            "example": 5432
          },
          "username": {
            "type": "string",
            "example": "myuser"
          },
          "password": {
            "type": "string",
            "example": "mypassword"
          },
          "database": {
            "type": "string",
            "example": "mydb"
          },
          "schema": {
            "type": "string",
            "example": "public"
          }
        }
      },
      "AliasMap": {
        "type": "object",
        "description": "Optional human-readable display names for tables and columns. Supplied at database analysis time; used in generated descriptions.\n",
        "properties": {
          "tables": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "orders": "Customer Orders"
            }
          },
          "columns": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            },
            "example": {
              "orders": {
                "order_date": "Order Date"
              }
            }
          }
        }
      },
      "AnalyzeDatabaseRequest": {
        "type": "object",
        "required": [
          "connectionInfo"
        ],
        "properties": {
          "connectionInfo": {
            "$ref": "#/components/schemas/PostgresConnectionInfo"
          },
          "aliasMap": {
            "$ref": "#/components/schemas/AliasMap"
          },
          "languageCode": {
            "type": "string",
            "description": "BCP 47 language code for error messages. Defaults to \"en\".",
            "example": "en"
          }
        }
      },
...
    }
  }
}